+++ aim/engine/aim_repo.py
@@ -4,6 +4,7 @@ import json
 import zipfile
 import re
 import time
+import base64
 
 from aim.__version__ import __version__ as aim_version
 from aim.engine.configs import *
@@ -55,6 +56,8 @@ class AimRepo:
             return AIM_CORR_DIR_NAME
         elif cat[0] == 'hyperparameters':
             return AIM_PARAMS_DIR_NAME
+        elif cat[0] == 'stats':
+            return AIM_STATS_DIR_NAME
 
     @staticmethod
     def archive_dir(zip_path, dir_path):
@@ -71,6 +74,8 @@ class AimRepo:
         self._config = {}
         self.path = os.path.join(path, AIM_REPO_NAME)
         self.config_path = os.path.join(self.path, AIM_CONFIG_FILE_NAME)
+        self.hash = base64.b64encode(path.encode('utf-8')).decode('utf-8')
+        self.name = path.split(os.sep)[-1]
 
         if self.config:
             self.branch = self.config.get('active_branch')
@@ -201,7 +206,7 @@ class AimRepo:
         meta_file.write(json.dumps(meta_file_content))
         meta_file.close()
 
-    def store_dir(self, name, data={}):
+    def store_dir(self, name, cat, data={}):
         """
         Creates a new directory inside repo and returns it's relative path
         """
@@ -216,6 +221,7 @@ class AimRepo:
         self.update_meta_file(name, {
             'name': name,
             'type': 'dir',
+            'cat': cat,
             'data': data,
             'data_path': dir_rel_path,
         })