+++ aim/sdk/artifacts/__init__.py
@@ -5,3 +5,4 @@ from aim.sdk.artifacts.model import *
 from aim.sdk.artifacts.correlation import *
 from aim.sdk.artifacts.distribution import *
 from aim.sdk.artifacts.hyperparameters import *
+from aim.sdk.artifacts.stats import *