+++ aim/sdk/track.py
@@ -5,10 +5,10 @@ import aim.logger
 
 
 def track(name: str, *args, **kwargs):
-
     if name not in globals():
         print('Aim cannot track: \'{0}\''.format(name))
         return
+
     # Get corresponding class
     obj = globals()[name]